+++ src/utils/formattedDiscountPrice.ts
@@ -10,8 +10,9 @@ export default function formattedDiscountPrice({
   discountType,
 }: FormattedDiscountPriceProps) {
   if (discountType === '할인율') {
-    // const stringValue = price.toString();
-    const numericValue = parseInt(price.replace(/,/g, ''), 10);
+    console.log(typeof price);
+    const stringValue = price.toString();
+    const numericValue = parseInt(stringValue.replace(/,/g, ''), 10);
     const returnValue = Math.round(numericValue * (1 - discount / 100));
     return returnValue;
   } else {
